requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pypdf==4.2.0
google-generativeai==0.8.0
python-multipart==0.0.9
pydantic-settings
python-dotenv
